--- Golang/EventSub/README.md
@@ -1,3 +1,0 @@
-**Required version** `1.18.0`
-
-**Twitch EventSub Broker**
