+++ Golang/cmd/EventSub/README.md
@@ -1,0 +1,1 @@
+**Twitch EventSub Broker**
