+++ Golang/cmd/EventSub/twitch_eventsub/eventsub_test.go
@@ -1,4 +1,4 @@
-package twitch
+package twitch_eventsub
 
 import (
 	"testing"
